Cast COUNT to int for Dapper CourseRow mapping
PostgreSQL COUNT returns bigint (Int64); CourseRow expects int (Int32).
Cast with ::int so Dapper can match the record constructor.

+++ backend/ShareMarketLMS.Api/Controllers/CoursesController.cs
@@ -27,8 +27,8 @@ public class CoursesController(AppDbContext db, DapperContext dapper) : Controll
 
         const string sql = """
             SELECT c."Id", c."Slug", c."Title", c."Description", c."Category",
-                   COUNT(l."Id") AS "TotalLessons",
-                   COUNT(lp."Id") AS "CompletedLessons"
+                   COUNT(l."Id")::int AS "TotalLessons",
+                   COUNT(lp."Id")::int AS "CompletedLessons"
             FROM "Courses" c
             LEFT JOIN "Modules" m ON m."CourseId" = c."Id"
             LEFT JOIN "Lessons" l ON l."ModuleId" = m."Id"
